Correctly quote ENV instructions
Docker uses the character defined by the #escape parser directive to
escape characters in strings. If we surround the environment variable
value with double-quotes, we only need to escape double-quotes and the
escape character itself in the value.

Using the OCaml quoting %S introduces wrong changes and doesn't work
when Docker escape character is changed.