Add support for Windows 10 LCU
Windows Containers cannot start an image for a newer version of Windows
and shouldn't attempt to start a container with a newer LCU (although
the container should physically start).

Adds support for appending the KB number for the LCU for a given month.
Microsoft only release Docker images for the official "update Tuesday"
LCU, and not any out-of-band LCUs, so the OOB LCUs are not included.
Since the KB number can differ by version, the LCU is exposed as a date
and is then mapped when given a Windows 10 release to a KB number to
append to the tag.