Ensure stray double-quotes don't get into PATH
If the registry value for Path ends with a backslash than \" ends up in
a command line which is interpreted as a double-quote character by the
somewhat messed up rules of command line parsing in Windows! The
"simplest" fix is to ensure this never happens. The actual fix would be
to add an extra backslash to the end of command lines when there's an
odd number of terminating backslashes - but try doing that before you've
got sed!