Refactor aliasing of distributions
Windows 11 / Server 2022 is code-named 21H2 within Microsoft, so adopt
that as the primary name for it in order to keep all Windows LTSC tags
as aliases.

Introduce Dockerfile_distro.distro which is the tags from
Dockerfile_distro.t which are not aliases (Stable, LTS, Latest, etc.).
This refactoring allows for some slightly more logical internal
structuring of the code, but the only externally visible change is the
return type of resolve_alias.