Dockerfile.mli: more documentation on mounts
The official docs are quite scarce on how these work, add a few more
details here to make it easier to use.

Most of these work with both Docker and Podman 4.x, some options are only
supported by Docker, and some options are parsed but do not work
on older versions of Podman.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>